Added function variable_range_data().
